RAG project/requirements.txt
streamlit
langchain-community
langchain
faiss-cpu
unstructured
unstructured[pdf]
transformers
sentence-transformers
huggingface_hub
torch
supabase
python-dotenv
pymupdf